Move "backtick identifiers" inner test class for unused private parameters
