+++ detekt-rules-style/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/style/UnusedPrivateMemberSpec.kt
@@ -769,17 +769,6 @@ class UnusedPrivateMemberSpec(val env: KotlinCoreEnvironment) {
     }
 
     @Nested
-    inner class `backtick identifiers - #5251` {
-        @Test
-        fun `does not report used backtick parameters`() {
-            val code = """
-                fun test(`foo bar`: Int) = `foo bar`
-            """.trimIndent()
-            assertThat(subject.compileAndLintWithContext(env, code)).isEmpty()
-        }
-    }
-
-    @Nested
     inner class `list get overloaded operator function - #3640` {
         @Test
         fun `report used private list get operator function - declared in a class - called by operator`() {
